+++ src/save_load.h
@@ -12,6 +12,5 @@ void save_error(Dungeon* dungeon);
 
 #include "floor.h"
 #include "save_load/file_version_0.h"
-#include "save_load/file_version_1.h"
 
 #endif